move download dno and gsp to script
