+++ pvsite_datamodel/write/data/dno.py
@@ -11,7 +11,6 @@ and we want to added the gsp as {gsp_id}|{gsp_nam} into the database
 """
 import logging
 import os
-import ssl
 
 import geopandas as gpd
 from shapely.geometry import Point
@@ -21,18 +20,6 @@ from pvsite_datamodel.write.data.utils import lat_lon_to_osgb
 logger = logging.getLogger(__name__)
 dir_path = os.path.dirname(os.path.realpath(__file__))
 dno_local_file = f"{dir_path}/dno"
-
-
-def download_dno():
-    """Download DNO data from NG."""
-
-    logger.debug("Getting dno file")
-    ssl._create_default_https_context = ssl._create_unverified_context
-    url = "https://data.nationalgrideso.com/backend/dataset/0e377f16-95e9-4c15-a1fc-49e06a39cfa0/resource/e96db306-aaa8-45be-aecd-65b34d38923a/download/dno_license_areas_20200506.geojson"  # noqa
-    dno_shapes = gpd.read_file(url)
-
-    logger.debug("Saving dno file")
-    dno_shapes.to_file(dno_local_file)
 
 
 def get_dno(latitude, longitude) -> dict:
